refactor: reorganize project modules
