+++ README.md
@@ -57,7 +57,7 @@ python gui/run_gui.py
 
 #### Command Line Interface
 ```python
-from src.agent_cell_phone import AgentCellPhone
+from services.agent_cell_phone import AgentCellPhone
 
 # Initialize agent
 acp = AgentCellPhone("agent-1")
@@ -106,25 +106,19 @@ acp.broadcast("Status update: All systems operational")
 
 ```
 Agent_CellPhone/
-├── launch.py                    # 🚀 Main launcher script
 ├── README.md                    # 📖 This file
 ├── requirements.txt             # 📦 Dependencies
-├── .gitignore                   # 🚫 Git ignore rules
 ├── src/                         # 🔧 Core system files
-│   ├── agent_cell_phone.py      # Core messaging system
-│   ├── inter_agent_framework.py # Inter-agent communication
+│   ├── core/                    # Core engine logic
+│   ├── gui/                     # GUI components
+│   ├── services/                # Background services
 │   └── main.py                  # Main system entry point
-├── gui/                         # 🖥️ GUI interfaces
-│   ├── dream_os_gui.py          # ✅ Modern PyQt GUI (main)
-│   ├── run_gui.py               # ✅ GUI launcher script
-│   └── agent_resume_web_gui.html # ✅ Web-based interface
 ├── tests/                       # 🧪 Test suite
 │   ├── test_harness.py          # Main test harness
 │   ├── test_8_agent_coordinates.py # 8-agent coordinate testing
 │   ├── test_inter_agent_framework.py # Framework testing
 │   ├── test_special_chars.py    # Special character testing
-│   ├── diagnostic_test.py       # Diagnostic testing tools
-│   └── coordinate_finder.py     # Coordinate mapping utility
+│   └── diagnostic_test.py       # Diagnostic testing tools
 ├── scripts/                     # 🔧 Utility scripts
 │   ├── agent_messenger.py       # Agent messaging utilities
 │   ├── agent_onboarding_sequence.py # Onboarding system
