+++ src/core/runtime/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime configuration and resources."""
