+++ src/core/runtime/config/cursor_agent_coords.json
