+++ src/core/runtime/config/templates/agent_modes.json
