+++ src/core/utils/coordinate_finder.py
