+++ src/gui/components/agent_panel.py
