+++ src/gui/components/onboarding_integration.py
