+++ src/gui/components/splash_screen.py
