+++ src/gui/logo.png
