+++ src/gui/run_two_agent_gui.py
@@ -6,11 +6,6 @@ Launcher script for the horizontal 2-agent GUI.
 """
 
 import sys
-import os
-
-# Add the project root to the path
-project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.insert(0, project_root)
 
 def main():
     """Launch the two agent horizontal GUI."""
@@ -28,4 +23,4 @@ def main():
         sys.exit(1)
 
 if __name__ == "__main__":
-    main() 
+    main()
