+++ src/services/__init__.py
@@ -1,0 +1,1 @@
+"""Background services and integration utilities."""
